entitysnapshot manager is moved to base class.
