+++ src/Abp.ZeroCore.EntityFrameworkCore/EntityHistory/EntitySnapshotManager.cs
@@ -14,7 +14,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Abp.EntityHistory
 {
-    public class EntitySnapshotManager : IEntitySnapshotManager, ITransientDependency
+    public class EntitySnapshotManager : EntitySnapshotManagerBase
     {
         private readonly IRepository<EntityChange, long> _entityChangeRepository;
 
@@ -22,71 +22,22 @@ namespace Abp.EntityHistory
         {
             _entityChangeRepository = entityChangeRepository;
         }
-        public async Task<EntityHistorySnapshot> GetSnapshotAsync<TEntity>(int id, DateTime snapshotTime) where TEntity : class, IEntity<int>
+
+        protected override async Task<TEntity> GetEntityById<TEntity, TPrimaryKey>(TPrimaryKey id)
         {
-            return await GetSnapshotAsync<TEntity, int>(id, snapshotTime);
+            return await _entityChangeRepository.GetDbContext()
+                .Set<TEntity>().AsQueryable().FirstOrDefaultAsync(base.CreateEqualityExpressionForId<TEntity, TPrimaryKey>(id));
         }
-        public async Task<EntityHistorySnapshot> GetSnapshotAsync<TEntity, TPrimaryKey>(TPrimaryKey id, DateTime snapshotTime) where TEntity : class, IEntity<TPrimaryKey>
+
+        protected override IQueryable<EntityChange> GetEntityChanges<TEntity, TPrimaryKey>(TPrimaryKey id, DateTime snapshotTime)
         {
-            var entity = _entityChangeRepository.GetDbContext()
-                .Set<TEntity>().AsQueryable().FirstOrDefault(x => x.Id.Equals(id));
+            string fullName = typeof(TEntity).FullName;
+            var idJson = id.ToJsonString();
 
-            var snapshotPropertiesDictionary = new Dictionary<string, string>();
-            var propertyChangesStackTreeDictionary = new Dictionary<string, string>();
-
-            if (entity != null)
-            {
-                string fullName = typeof(TEntity).FullName;
-                var idJson = id.ToJsonString();
-
-                var changes = await _entityChangeRepository.GetAll()//select all changes which created after snapshot time 
-                    .Where(x => x.EntityTypeFullName == fullName && x.EntityId == idJson && x.ChangeTime > snapshotTime && x.ChangeType != EntityChangeType.Created)
-                    .OrderByDescending(x => x.ChangeTime)
-                    .Select(x => new { x.ChangeType, x.PropertyChanges }).ToListAsync();
-
-                //and revoke all changes
-
-                foreach (var change in changes)// desc ordered changes
-                {
-                    foreach (var entityPropertyChange in change.PropertyChanges)
-                    {
-                        if (snapshotPropertiesDictionary.ContainsKey(entityPropertyChange.PropertyName))
-                        {
-                            snapshotPropertiesDictionary[entityPropertyChange.PropertyName] = entityPropertyChange.OriginalValue;//set back to orginal value
-                        }
-                        else
-                        {
-                            snapshotPropertiesDictionary.Add(entityPropertyChange.PropertyName, entityPropertyChange.OriginalValue);//set back to orginal value
-                        }
-
-                        //create change stack tree
-                        if (propertyChangesStackTreeDictionary.ContainsKey(entityPropertyChange.PropertyName))
-                        {
-                            propertyChangesStackTreeDictionary[entityPropertyChange.PropertyName] += " -> " + entityPropertyChange.OriginalValue;
-                        }
-                        else
-                        {
-                            string propertyCurrentValue = "PropertyNotExist";
-
-                            var propertyInfo = typeof(TEntity).GetProperty(entityPropertyChange.PropertyName);
-                            if (propertyInfo != null)
-                            {
-                                var val = propertyInfo.GetValue(entity);
-                                if (val == null)
-                                {
-                                    propertyCurrentValue = "null";
-                                }
-                                else
-                                {
-                                    propertyCurrentValue = val.ToJsonString();
-                                }
-                            }
-                            propertyChangesStackTreeDictionary.Add(entityPropertyChange.PropertyName, propertyCurrentValue + " -> " + entityPropertyChange.OriginalValue);
-                        }
-                    }
-                }
-            }
-            return new EntityHistorySnapshot(snapshotPropertiesDictionary, propertyChangesStackTreeDictionary);
+            return _entityChangeRepository.GetAll() //select all changes which created after snapshot time 
+                .Where(x => x.EntityTypeFullName == fullName && x.EntityId == idJson && x.ChangeTime > snapshotTime &&
+                            x.ChangeType != EntityChangeType.Created)
+                .OrderByDescending(x => x.ChangeTime);
         }
     }
 }
